Rearrange structure of lua_env
