+++ src/character.cpp
@@ -1163,7 +1163,7 @@ int chara_create_internal()
     cdata[rc].index = rc;
     initialize_character();
 
-    lua::lua->on_chara_creation(cdata[rc]);
+    lua::lua->create_chara_handle_run_callbacks(cdata[rc]);
 
     rtval = rc;
     return 1;
@@ -2069,7 +2069,7 @@ void chara_killed(character& chara)
         // This character slot is invalid, and can be overwritten by
         // newly created characters at any time. Run any Lua callbacks
         // to clean up character things.
-        lua::lua->on_chara_removal(chara);
+        lua::lua->remove_chara_handle_run_callbacks(chara);
     }
     else if (chara.state == 2 || chara.state == 4 || chara.state == 6)
     {
@@ -2092,7 +2092,7 @@ void chara_delete(int cc)
         // currently valid. If the state were 0, then chara_killed
         // would have been called to run the chara removal handler for
         // the Lua state. We'll have to run it now.
-        lua::lua->on_chara_removal(cdata[cc]);
+        lua::lua->remove_chara_handle_run_callbacks(cdata[cc]);
     }
     else
     {
@@ -2205,7 +2205,7 @@ int chara_relocate(int prm_784, int prm_785, int prm_786)
     }
 
     // TODO handle transferring through Lua robustly
-    // lua::lua->on_chara_removal(cdata[prm_784]);
+    // lua::lua->remove_chara_handle_run_callbacks(cdata[prm_784]);
 
     sdata.copy(tc_at_m125, prm_784);
     sdata.clear(prm_784);
@@ -2282,7 +2282,7 @@ int chara_relocate(int prm_784, int prm_785, int prm_786)
     chara_refresh(tc_at_m125);
 
     // TODO handle transferring through Lua robustly
-    // lua::lua->on_chara_creation(cdata[tc_at_m125]);
+    // lua::lua->create_chara_handle_run_callbacks(cdata[tc_at_m125]);
 
     if (tc_at_m125 < 57)
     {
