+++ src/i18n.hpp
@@ -12,7 +12,7 @@
 #include "hcl.hpp"
 #include "item.hpp"
 #include "log.hpp"
-#include "lua_env/lua_api.hpp"
+#include "lua_env/api_manager.hpp"
 #include "macro.hpp"
 #include "optional.hpp"
