+++ src/initialize_map.cpp
@@ -72,7 +72,7 @@ label_17401:
                 if (cdata[cnt].state == 9)
                 {
                     cdata[cnt].state = 1;
-                    lua::lua->on_chara_loaded(cdata[cnt]);
+                    lua::lua->get_handle_manager().create_character_handle(cdata[cnt]);
                 }
             }
         }
@@ -2707,7 +2707,7 @@ label_1742_internal:
         }
         rc = cnt;
         cdata[rc].state = 1;
-        lua::lua->on_chara_loaded(cdata[rc]);
+        lua::lua->get_handle_manager().create_character_handle(cdata[rc]);
         if (cdata[cnt].is_contracting() == 1)
         {
             cxinit = cdata[0].position.x;
