+++ src/lua_env/lua_env.cpp
@@ -1,23 +1,11 @@
 #include "lua_env.hpp"
-#include <map>
-#include <sstream>
-#include <vector>
-#include "../character.hpp"
-#include "../config.hpp"
-#include "../elona.hpp"
-#include "../filesystem.hpp"
-#include "../log.hpp"
-#include "../variables.hpp"
-#include "event_manager.hpp"
 
 namespace elona
 {
 namespace lua
 {
 
-std::unique_ptr<lua_env> lua;
-
-lua_env::lua_env()
+void lua_env::initialize()
 {
     lua = std::make_shared<sol::state>();
     lua->open_libraries(
@@ -34,413 +22,15 @@ lua_env::lua_env()
     std::string normalized = filesystem::to_forward_slashes(exe_path);
     lua->safe_script(u8"package.path = \""s + normalized + u8"/?.lua;\"..package.path"s);
 
-    // Create dummy Store table to prevent crashes on access from
-    // state outside of an environment
-    lua->create_named_table("Store");
-
     // Make sure the API environment is initialized first so any
     // dependent managers can add new internal C++ methods to it (like
     // the event manager registering Elona.Event)
     api_mgr = std::make_unique<api_manager>(this);
-
     event_mgr = std::make_unique<event_manager>(this);
-    event_manager::init(*this);
-
+    mod_mgr = std::make_unique<mod_manager>(this);
     handle_mgr = std::make_unique<handle_manager>(this);
-
     registry_mgr = std::make_unique<registry_manager>(this);
-    registry_manager::init(*this);
 }
-
-api_manager& lua_env::get_api_manager()
-{
-    return *api_mgr;
-}
-
-event_manager& lua_env::get_event_manager()
-{
-    return *event_mgr;
-}
-
-handle_manager& lua_env::get_handle_manager()
-{
-    return *handle_mgr;
-}
-
-registry_manager& lua_env::get_registry_manager()
-{
-    return *registry_mgr;
-}
-
-void report_error(sol::error err)
-{
-    std::string what = err.what();
-    ELONA_LOG(what);
-}
-
-
-// Handlers for existing characters/items being loaded from save archives
-void lua_env::on_chara_loaded(character& chara)
-{
-    if (chara.state == 0)
-    {
-        return;
-    }
-    handle_mgr->create_chara_handle(chara);
-}
-
-void lua_env::on_item_loaded(item& item)
-{
-    if (item.number == 0)
-    {
-        return;
-    }
-    handle_mgr->create_item_handle(item);
-}
-
-
-// Handlers for characters/items being unloaded when another save archive is
-// loaded
-void lua_env::on_chara_unloaded(character& chara)
-{
-    if (chara.state == 0)
-    {
-        return;
-    }
-    handle_mgr->remove_chara_handle(chara);
-}
-
-void lua_env::on_item_unloaded(item& item)
-{
-    if (item.number == 0)
-    {
-        return;
-    }
-    handle_mgr->remove_item_handle(item);
-}
-
-
-// Handlers for brand-new instances of characters/objects being created
-void lua_env::on_chara_creation(character& chara)
-{
-    assert(chara.state != 0);
-    handle_mgr->create_chara_handle(chara);
-
-    auto handle = handle_mgr->get_chara_handle(chara);
-    assert(handle != sol::lua_nil);
-    event_mgr->run_callbacks<event_kind_t::character_created>(handle);
-}
-
-void lua_env::on_item_creation(item& item)
-{
-    assert(item.number != 0);
-    handle_mgr->create_item_handle(item);
-
-    auto handle = handle_mgr->get_item_handle(item);
-    assert(handle != sol::lua_nil);
-    event_mgr->run_callbacks<event_kind_t::item_created>(handle);
-}
-
-
-// Handlers for invalidation of characters/items (character death, item count is
-// 0)
-void lua_env::on_chara_removal(character& chara)
-{
-    auto handle = handle_mgr->get_chara_handle(chara);
-    event_mgr->run_callbacks<event_kind_t::character_removed>(handle);
-
-    handle_mgr->remove_chara_handle(chara);
-}
-
-void lua_env::on_item_removal(item& item)
-{
-    auto handle = handle_mgr->get_item_handle(item);
-    event_mgr->run_callbacks<event_kind_t::item_removed>(handle);
-
-    handle_mgr->remove_item_handle(item);
-}
-
-void lua_env::clear_map_local_data()
-{
-    for (auto&& pair : mods)
-    {
-        auto& mod = pair.second;
-        lua->safe_script(R"(
-local function clear(t)
-    for key, _ in pairs(t) do
-        t[key] = nil
-    end
-end
-if Store then
-    clear(Store.map_local)
-end
-)", mod->env);
-    }
-    handle_mgr->clear_map_local_handles();
-}
-
-
-void lua_env::load_mod(mod_info& mod)
-{
-    setup_and_lock_mod_globals(mod);
-
-    if (mod.path) // Skip initializing mods not created from files.
-    {
-        auto result = this->lua->safe_script_file(
-            filesystem::make_preferred_path_in_utf8(*mod.path / u8"init.lua"s),
-            mod.env);
-
-        // Add the API table returned by the mod's init.lua, if one
-        // was returned.
-        if (result.valid())
-        {
-            sol::optional<sol::object> object = result.get<sol::object>();
-            if (object && object->is<sol::table>())
-            {
-                sol::table api_table = object->as<sol::table>();
-                api_mgr->add_api(mod.name, api_table);
-            }
-        }
-        else
-        {
-            sol::error err = result;
-            report_error(err);
-            throw std::runtime_error("Failed initializing mod "s + mod.name);
-        }
-    }
-}
-
-void lua_env::scan_all_mods(const fs::path& mods_dir)
-{
-    if (stage != mod_loading_stage_t::not_started
-          && stage != mod_loading_stage_t::scan_finished)
-    {
-        throw std::runtime_error("Mods have already been scanned!");
-    }
-
-    const std::string init_script = "init.lua";
-
-    // TODO: [dependency management] order mods and always load core first.
-    for (const auto& entry :
-         filesystem::dir_entries{mods_dir, filesystem::dir_entries::type::dir})
-    {
-        if (fs::exists(entry.path() / init_script))
-        {
-            const std::string mod_name = entry.path().filename().string();
-            ELONA_LOG("Found mod " << mod_name);
-
-            // TODO verify the mod name is alphanumeric only.
-
-            if (mod_name == "script")
-            {
-                throw std::runtime_error("\"script\" is a reserved mod name.");
-            }
-
-            std::unique_ptr<mod_info> info =
-                std::make_unique<mod_info>(mod_name, entry.path(), get_state());
-            this->mods.emplace(mod_name, std::move(info));
-        }
-    }
-    stage = mod_loading_stage_t::scan_finished;
-}
-
-void lua_env::load_core_mod()
-{
-    if (stage == mod_loading_stage_t::not_started)
-    {
-        throw std::runtime_error("Mods haven't been scanned yet!");
-    }
-
-    auto val = this->mods.find("core");
-    if (val == this->mods.end())
-    {
-        throw std::runtime_error(
-            "Core mod was not found. Does \"mods/core\" exist?");
-    }
-
-    // Add special API tables from data/lua to the core mod. The core
-    // API table will be modified in-place by the Lua API code.
-    api_mgr->load_core(*this);
-    stage = mod_loading_stage_t::core_mod_loaded;
-}
-
-void lua_env::load_all_mods()
-{
-    if (stage != mod_loading_stage_t::core_mod_loaded)
-    {
-        throw std::runtime_error("Core mod wasn't loaded!");
-    }
-    for (auto& pair : this->mods)
-    {
-        auto& mod = pair.second;
-        if (mod->name == "script")
-        {
-            // TODO warn about reserved mod names.
-            continue;
-        }
-        else
-        {
-            load_mod(*mod);
-        }
-        ELONA_LOG("Loaded mod " << mod->name);
-    }
-
-    event_mgr->run_callbacks<event_kind_t::all_mods_loaded>();
-    registry_mgr->register_functions();
-
-    stage = mod_loading_stage_t::all_mods_loaded;
-}
-
-void lua_env::run_startup_script(const std::string& name)
-{
-    if (stage < mod_loading_stage_t::core_mod_loaded)
-    {
-        throw std::runtime_error("Core mod wasn't loaded!");
-    }
-    if (this->mods.find(name) != this->mods.end())
-    {
-        throw std::runtime_error("Startup script was already run.");
-    }
-
-    std::unique_ptr<mod_info> script_mod =
-        std::make_unique<mod_info>("script", none, get_state());
-    setup_and_lock_mod_globals(*script_mod);
-
-    lua->safe_script_file(
-        filesystem::make_preferred_path_in_utf8(
-            filesystem::dir::data() / "script"s / name),
-        script_mod->env);
-
-    ELONA_LOG("Loaded startup script " << name);
-    txtef(8);
-    txt(lang(
-        u8"スクリプト"s + config::instance().startup_script
-            + u8"が読み込まれました。"s,
-        u8"Loaded script "s + config::instance().startup_script + u8". "s));
-    txtnew();
-
-    this->mods.emplace("script", std::move(script_mod));
-}
-
-void lua_env::clear_mod_stores()
-{
-    for (auto&& pair : mods)
-    {
-        auto& mod = pair.second;
-        lua->safe_script(R"(
-local function clear(t)
-    for key, _ in pairs(t) do
-        t[key] = nil
-    end
-end
-if Store then
-    clear(Store.map_local)
-    clear(Store.global)
-end
-)", mod->env);
-    }
-}
-
-void lua_env::reload()
-{
-    clear(); // Unload character/item handles while they're still available.
-    get_state()->set("_IS_TEST", config::instance().is_test);
-    scan_all_mods(filesystem::dir::mods());
-    load_core_mod();
-    load_all_mods();
-}
-
-// Callback to be called in Lua for preventing write access to unknown
-// globals.
-int deny(
-    sol::table table,
-    sol::object key,
-    sol::object value,
-    sol::this_state ts)
-{
-    UNUSED(table);
-    UNUSED(value);
-
-    std::stringstream ss;
-    if (key.is<std::string>())
-    {
-        ss << "Cannot assign to the global variable \"" << key.as<std::string>()
-           << "\". ";
-    }
-    else
-    {
-        ss << "An attempt was made to assign to a global variable. ";
-    }
-
-    ss << "Please prefix the assignment with \"local\" to make it a local "
-          "variable.";
-
-    lua_State* L = ts;
-    return luaL_error(L, ss.str().c_str());
-}
-
-void lua_env::bind_store(sol::state& lua, mod_info& mod, sol::table& table)
-{
-    sol::table Store = lua.create_table();
-    sol::table metatable = lua.create_table();
-
-    // Bind Store.global and Store.map_local.
-    metatable["global"] = mod.store_global;
-    metatable["map_local"] = mod.store_local;
-
-    // Prevent creating new variables in the Store table.
-    metatable[sol::meta_function::new_index] = deny;
-    metatable[sol::meta_function::index] = metatable;
-
-    Store[sol::metatable_key] = metatable;
-    table["Store"] = Store;
-}
-
-void lua_env::setup_mod_globals(mod_info& mod, sol::table& table)
-{
-    // Create the globals "Elona" and "Store" for this mod's
-    // environment.
-    bind_store(*lua, mod, table);
-    table["Elona"] = api_mgr->bind(*this);
-    table["_MOD_NAME"] = mod.name;
-
-    // Add a list of whitelisted standard library functions to the
-    // environment.
-    setup_sandbox(*this->lua, table);
-
-    // Add a custom version of "require" for use within mods. (Not
-    // added for scripts/console environment)
-    if (mod.path)
-    {
-        auto state = this->get_state();
-        fs::path mod_path = *mod.path;
-        table["require"] =
-            [state, mod_path](const std::string& path, sol::this_environment this_env)
-            {
-                sol::environment env = this_env;
-                const fs::path full_path = mod_path / (path + ".lua");
-                return state->script_file(full_path.string(), env);
-            };
-    }
-}
-
-void lua_env::setup_and_lock_mod_globals(mod_info& mod)
-{
-    sol::table env_metatable = lua->create_table_with();
-
-    // Globals have to be set on the metatable, not the mod's
-    // environment itself.
-    setup_mod_globals(mod, env_metatable);
-
-    // Prevent writing of new globals.
-    env_metatable[sol::meta_function::new_index] = deny;
-    env_metatable[sol::meta_function::index] = env_metatable;
-
-    mod.env[sol::metatable_key] = env_metatable;
-}
-
-
-// For testing use
 
 void lua_env::clear()
 {
@@ -448,7 +38,7 @@ void lua_env::clear()
     {
         if (inv[i].number != 0)
         {
-            on_item_unloaded(inv[i]);
+            handle_mgr->remove_item_handle(inv[i]);
         }
     }
 
@@ -456,88 +46,26 @@ void lua_env::clear()
     {
         if (cdata[i].state != 0)
         {
-            on_chara_unloaded(cdata[i]);
+            handle_mgr->remove_chara_handle(cdata[i]);
         }
     }
+
     event_mgr->clear();
-    clear_mod_stores();
+    mod_mgr->clear_mod_stores();
     lua->collect_garbage();
+
+    // TODO needed?
     // mods.clear();
     // stage = mod_loading_stage_t::not_started;
 }
 
-void lua_env::load_mod_from_script(
-    const std::string& name,
-    const std::string& script,
-    bool readonly)
+void lua_env::reload()
 {
-    if (stage < mod_loading_stage_t::core_mod_loaded)
-    {
-        throw std::runtime_error("Core mod wasn't loaded!");
-    }
-    {
-        auto val = mods.find(name);
-        if (val != mods.end())
-            throw std::runtime_error(
-                "Mod "s + name + " was already initialized."s);
-    }
+    clear(); // Unload character/item handles while they're still available.
+    get_state()->set("_IS_TEST", config::instance().is_test);
 
-    std::unique_ptr<mod_info> info =
-        std::make_unique<mod_info>(name, none, get_state());
-
-    if (readonly)
-    {
-        setup_and_lock_mod_globals(*info);
-    }
-    else
-    {
-        // Set the globals directly on the environment table for testing use.
-        setup_mod_globals(*info, info->env);
-    }
-
-    // Add the API table returned by the mod's initialization script,
-    // if one was returned.
-    auto result = this->lua->safe_script(script, info->env);
-    if (result.valid())
-    {
-        sol::optional<sol::object> object = result.get<sol::object>();
-        if (object && object->is<sol::table>())
-        {
-            sol::table api_table = object->as<sol::table>();
-            api_mgr->add_api(name, api_table);
-        }
-    }
-    else
-    {
-        sol::error err = result;
-        report_error(err);
-        throw std::runtime_error("Failed initializing mod "s + name);
-    }
-
-    this->mods[name] = std::move(info);
-}
-
-void lua_env::run_in_mod(const std::string& name, const std::string& script)
-{
-    auto val = mods.find(name);
-    if (val == mods.end())
-    {
-        throw std::runtime_error("No such mod "s + name + "."s);
-    }
-
-    // Prevent outputting things in the test log on errors, but still
-    // throw if the result is invalid so tests can catch it.
-    auto ignore_handler = [](lua_State*, sol::protected_function_result pfr) {
-        return pfr;
-    };
-
-    auto result = this->lua->script(script, val->second->env, ignore_handler);
-
-    if (!result.valid())
-    {
-        sol::error err = result;
-        throw err;
-    }
+    // TODO needed?
+    //mod_mgr->load_mods(filesystem::dir::mods());
 }
 
 } // namespace lua
