+++ src/lua_env/registry_manager.cpp
